requirements.txt
xgboost>=2.0.0
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.3.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
